requirements.txt
asgiref==3.8.1
Django==5.1.2
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
gunicorn==23.0.0
packaging==24.1
psycopg2-binary==2.9.9
PyJWT==2.9.0
sqlparse==0.5.1
whitenoise==6.7.0

python-dotenv==0.21.1


